requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
chromadb==0.5.18
sentence-transformers==3.3.1
cryptography==44.0.0
pypdf2==3.0.1
pdfplumber==0.11.4
jinja2==3.1.4
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
python-docx==1.2.0
python-pptx==1.0.2
reportlab==4.5.0
